add networkManager and use in publish script
